Move to package io.awspring.cloud.samples
