+++ spring-cloud-aws-samples/spring-cloud-aws-parameter-store-sample/app/src/main/java/io/awspring/cloud/samples/parameterstore/SpringCloudAwsParameterStoreSample.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.awspring.cloud.parameterstore.sample;
+package io.awspring.cloud.samples.parameterstore;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
