+++ spring-cloud-aws-samples/spring-cloud-aws-parameter-store-sample/infrastructure/src/main/java/io/awspring/cloud/samples/parameterstore/infrastructure/InfrastructureApp.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.awspring.cloud.parameterstore.sample.infrastructure;
+package io.awspring.cloud.samples.parameterstore.infrastructure;
 
 import software.amazon.awscdk.App;
 import software.amazon.awscdk.StackProps;
