+++ spring-cloud-aws-samples/spring-cloud-aws-parameter-store-sample/infrastructure/src/main/java/io/awspring/cloud/samples/parameterstore/infrastructure/InfrastructureStack.java
@@ -13,19 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.awspring.cloud.secretsmanager.sample.infrastructure;
+package io.awspring.cloud.samples.parameterstore.infrastructure;
 
-import software.amazon.awscdk.App;
+import software.amazon.awscdk.Stack;
 import software.amazon.awscdk.StackProps;
+import software.amazon.awscdk.services.ssm.StringParameter;
+import software.constructs.Construct;
 
-public class InfrastructureApp {
+public class InfrastructureStack extends Stack {
 
-	public static void main(final String[] args) {
-		App app = new App();
+	public InfrastructureStack(final Construct scope, final String id) {
+		this(scope, id, null);
+	}
 
-		new InfrastructureStack(app, "InfrastructureStack", StackProps.builder().build());
+	public InfrastructureStack(final Construct scope, final String id, final StackProps props) {
+		super(scope, id, props);
 
-		app.synth();
+		StringParameter.Builder.create(this, "Parameter").parameterName("/config/spring/message")
+				.stringValue("Spring-cloud-aws value!").build();
 	}
 
 }
