+++ spring-cloud-aws-samples/spring-cloud-aws-s3-sample/app/src/main/java/io/awspring/cloud/samples/s3/SpringCloudAwsS3Sample.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.awspring.cloud.s3.sample;
+package io.awspring.cloud.samples.s3;
 
 import java.io.IOException;
 import java.io.OutputStream;
