+++ spring-cloud-aws-samples/spring-cloud-aws-s3-sample/infrastructure/pom.xml
@@ -22,7 +22,7 @@
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>exec-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>io.awspring.cloud.s3.sample.infrastructure.InfrastructureApp</mainClass>
+                    <mainClass>io.awspring.cloud.samples.s3.infrastructure.InfrastructureApp</mainClass>
                 </configuration>
             </plugin>
         </plugins>
