+++ spring-cloud-aws-samples/spring-cloud-aws-s3-sample/infrastructure/src/main/java/io/awspring/cloud/samples/s3/infrastructure/InfrastructureStack.java
@@ -13,12 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.awspring.cloud.secretsmanager.sample.infrastructure;
+package io.awspring.cloud.samples.s3.infrastructure;
 
+import java.util.Arrays;
 import software.amazon.awscdk.Stack;
 import software.amazon.awscdk.StackProps;
-import software.amazon.awscdk.services.secretsmanager.Secret;
-import software.amazon.awscdk.services.secretsmanager.SecretStringGenerator;
+import software.amazon.awscdk.services.s3.Bucket;
+import software.amazon.awscdk.services.s3.deployment.BucketDeployment;
+import software.amazon.awscdk.services.s3.deployment.Source;
 import software.constructs.Construct;
 
 public class InfrastructureStack extends Stack {
@@ -30,11 +32,12 @@ public class InfrastructureStack extends Stack {
 	public InfrastructureStack(final Construct scope, final String id, final StackProps props) {
 		super(scope, id, props);
 
-		SecretStringGenerator secretStringGenerator = SecretStringGenerator.builder().generateStringKey("password")
-				.secretStringTemplate("{}").build();
+		Bucket bucket1 = Bucket.Builder.create(this, "bucket1").bucketName("spring-cloud-aws-sample-bucket1").build();
 
-		Secret.Builder.create(this, "Secret").secretName("/secrets/spring-cloud-aws-sample-app")
-				.generateSecretString(secretStringGenerator).build();
+		BucketDeployment.Builder.create(this, "bucketDeployment").destinationBucket(bucket1)
+				.sources(Arrays.asList(Source.data("my-file.txt", "my file content"),
+						Source.data("test-file.txt", "test file content")))
+				.build();
 	}
 
 }
