+++ spring-cloud-aws-samples/spring-cloud-aws-secrets-manager-sample/app/pom.xml
@@ -7,7 +7,9 @@
 		<artifactId>spring-cloud-aws-samples-secrets-manager</artifactId>
 		<version>3.0.0-SNAPSHOT</version>
 	</parent>
- 	<artifactId>secrets-manager-sample</artifactId>
+
+ 	<artifactId>secrets-manager-sample-app</artifactId>
+	<name>Spring Cloud AWS Secrets Manager Sample App</name>
 
 	<dependencies>
 		<dependency>
