+++ spring-cloud-aws-samples/spring-cloud-aws-secrets-manager-sample/infrastructure/pom.xml
@@ -22,7 +22,7 @@
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>exec-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>io.awspring.cloud.secretsmanager.sample.infrastructure.InfrastructureApp</mainClass>
+                    <mainClass>io.awspring.cloud.samples.secretsmanager.infrastructure.InfrastructureApp</mainClass>
                 </configuration>
             </plugin>
         </plugins>
